ISSUE-298 Move CassandraRateLimitationPlanRepository and CassandraRateLimitingPlanUserRepository to the same package
