+++ tmail-backend/rate-limiter/rate-limiter-cassandra/src/main/scala/com/linagora/tmail/rate/limiter/api/cassandra/CassandraRateLimitationPlanRepository.scala
@@ -1,6 +1,7 @@
-package com.linagora.tmail.rate.limiter.cassandra
+package com.linagora.tmail.rate.limiter.api.cassandra
 
 import com.google.common.base.Preconditions
+import com.linagora.tmail.rate.limiter.api.cassandra.dao.{CassandraRateLimitPlanDAO, RateLimitingPlanEntry}
 import com.linagora.tmail.rate.limiter.api.{RateLimitationPlanRepository, RateLimitingPlan, RateLimitingPlanCreateRequest, RateLimitingPlanId, RateLimitingPlanName, RateLimitingPlanNotFoundException, RateLimitingPlanResetRequest}
 import javax.inject.Inject
 import org.apache.james.util.ReactorUtils.DEFAULT_CONCURRENCY
