+++ tmail-backend/rate-limiter/rate-limiter-cassandra/src/main/scala/com/linagora/tmail/rate/limiter/api/cassandra/dao/CassandraRateLimitPlanDAO.scala
@@ -1,19 +1,20 @@
-package com.linagora.tmail.rate.limiter.cassandra
+package com.linagora.tmail.rate.limiter.api.cassandra.dao
+
+import java.time.Duration
+import java.util
+import java.util.UUID
 
 import com.datastax.driver.core.DataType.{bigint, text}
 import com.datastax.driver.core.querybuilder.QueryBuilder
 import com.datastax.driver.core.querybuilder.QueryBuilder.{bindMarker, insertInto, select}
 import com.datastax.driver.core.{DataType, PreparedStatement, Row, Session, TupleType, TupleValue}
+import com.linagora.tmail.rate.limiter.api.cassandra.table.CassandraRateLimitPlanHeaderEntry.{RATE_LIMITATION_DURATION_INDEX, RATE_LIMITATION_NAME_INDEX, RATE_LIMITS_INDEX}
+import com.linagora.tmail.rate.limiter.api.cassandra.table.CassandraRateLimitPlanTable.{OPERATION_LIMITATION_NAME, PLAN_ID, PLAN_NAME, RATE_LIMITATIONS, TABLE_NAME}
 import com.linagora.tmail.rate.limiter.api.{LimitTypes, OperationLimitations, RateLimitation, RateLimitingPlanCreateRequest, RateLimitingPlanId, RateLimitingPlanResetRequest}
-import com.linagora.tmail.rate.limiter.cassandra.CassandraRateLimitPlanHeaderEntry.{RATE_LIMITATION_DURATION_INDEX, RATE_LIMITATION_NAME_INDEX, RATE_LIMITS_INDEX}
-import com.linagora.tmail.rate.limiter.cassandra.CassandraRateLimitPlanTable.{OPERATION_LIMITATION_NAME, PLAN_ID, PLAN_NAME, RATE_LIMITATIONS, TABLE_NAME}
+import javax.inject.Inject
 import org.apache.james.backends.cassandra.utils.CassandraAsyncExecutor
 import reactor.core.scala.publisher.{SFlux, SMono}
 
-import java.time.Duration
-import java.util
-import java.util.UUID
-import javax.inject.Inject
 import scala.jdk.CollectionConverters._
 
 object RateLimitingPlanEntry {
